backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
anthropic
httpx
python-multipart
python-dotenv
websockets
google-genai
